fixing url
